#1 organisms/Main作成
