+++ react-quote-table/src/App.css
@@ -7,7 +7,5 @@ body{
     font-family: "Nasu-Regular";
 }
 
-.main{
-    margin: 100px 0;
-}
+
