+++ react-quote-table/src/components/organisms/Main/index.module.css
@@ -1,0 +1,3 @@
+.main{
+    margin: 100px 0;
+}
